split a file to support macOS
